fix delete house and fix layout service chosen

+++ app/controllers/houses_controller.rb
@@ -82,8 +82,10 @@ class HousesController < ApplicationController
       firebase = Firebase::Client.new(FIREBASE_URL, FIREBASE_SECRET)
       firebase.delete("#{upercase_house_name}")
       user = User.find_by_house_id(params[:id])
-      Reminder.where(user_id: user.id).delete_all
-      User.where(house_id: params[:id]).delete_all
+      if !user.nil?
+        Reminder.where(user_id: user.id).delete_all
+        User.where(house_id: params[:id]).delete_all
+      end
       house = House.delete(params[:id])
       flash[:notice] = I18n.t('mes.delete_success', name: I18n.t('house.house_name'))
       redirect_to houses_path
